Fix active learning restart 2 (#57)
* Added new tests of copying parameters on restart, changed how fb files are read in on restart and how ff files are copied during parameter optimization

* Altered checkArgs to properly check trainMdin and validMdin; reformatted

* Everything outside activeLearning looks fine; tests passing (outside activeLearning)

* Removed pdbs from active_learning; all tests passing

* Reformatted

* Cleaned up files being produced by optengine test cases

* ... Put back the files I actually needed

* Removed use of cpptraj

* Redo geometry collection on restart

* Found the last hideout of the pdbs

* Found and fixed bug for active learning restart

* reformatted

+++ tests/test_optengine.py
@@ -14,7 +14,7 @@ def monkeyGraph():
     pass
 
 
-def cleanOptDir(optdir):
+def cleanOptDir(optdir, removeResult=False):
     try:
         rmtree(os.path.join(optdir, "forcefield"))
     except:
@@ -23,16 +23,18 @@ def cleanOptDir(optdir):
         os.remove(os.path.join(optdir, "valid_0_initial.in"))
     if os.path.isfile(os.path.join(optdir, "setup_valid_initial.leap")):
         os.remove(os.path.join(optdir, "setup_valid_initial.leap"))
-    try:
-        rmtree(os.path.join(optdir, "result"))
-    except:
-        pass
+    if removeResult:
+        try:
+            rmtree(os.path.join(optdir, "result"))
+        except:
+            pass
     for f in os.listdir(optdir):
         if (
             f.startswith("prev")
             or f.endswith(".inpcrd")
             or f.endswith(".prmtop")
             or f == "leap.out"
+            or f == "leap.log"
         ):
             os.remove(os.path.join(optdir, f))
 
@@ -489,7 +491,7 @@ def test_optimizeForcefield0(monkeypatch):
         testLines = f.readlines()
     os.remove("opt_0.out")
     os.chdir("..")
-    cleanOptDir(options["optdir"])
+    cleanOptDir(options["optdir"], removeResult=True)
     assert filesFound
     assert checkUtils.checkLists(testMol2Lines, refMol2Lines)
     assert checkUtils.checkLists(testLines, refLines)
@@ -536,7 +538,7 @@ def test_optimizeForcefield1(monkeypatch):
         testLines = f.readlines()
     assert checkUtils.checkLists(testLines, refLines)
     os.chdir("..")
-    cleanOptDir(options["optdir"])
+    cleanOptDir(options["optdir"], removeResult=True)
 
 
 # Finished FB cycle
@@ -688,7 +690,7 @@ def test_respPriors(monkeypatch):
         "prev_dasa.mol2", os.path.join("ref", "dasa_priors.mol2")
     )
     os.chdir("..")
-    cleanOptDir(options["optdir"])
+    cleanOptDir(options["optdir"], removeResult=True)
 
 
 def test_restartResp():
@@ -803,7 +805,7 @@ def test_optimizeForcefield_multipleValids(monkeypatch):
             valid2 += 1
             os.remove(f)
     os.chdir("..")
-    cleanOptDir(options["optdir"])
+    cleanOptDir(options["optdir"], removeResult=True)
     assert valid1 == 3
     assert valid2 == 3
 
@@ -826,13 +828,13 @@ def test_determineRestart_multipleValids():
 
 
 def monkeyForcebalance(command):
-    with open("fb.log","a") as f:
+    with open("fb.log", "a") as f:
         f.write(command + "\n")
     inFile = command.split()[1]
-    outFileSplit = command.split()[3].replace(".out","").split("_")
+    outFileSplit = command.split()[3].replace(".out", "").split("_")
     if outFileSplit[0] == "opt":
         fbType = "opt"
-    else:   
+    else:
         if len(outFileSplit) == 2:
             fbType = "valid"
         else:
@@ -886,6 +888,7 @@ def setupFFdir(optdir):
     copyfile("dasa.mol2", os.path.join("forcefield", "initial_dasa.mol2"))
     os.chdir("..")
 
+
 # Restart from new FB cycle
 def test_restart1Params(monkeypatch):
     monkeypatch.setattr(optengine.OptEngine, "sortParams", monkeySortParams)
@@ -913,6 +916,7 @@ def test_restart1Params(monkeypatch):
         lines = f.readlines()
     os.remove("fb.log")
     rmtree("forcefield")
+    cleanOptDir(".")
     os.chdir("..")
     assert len(lines) == 4
 
@@ -945,6 +949,7 @@ def test_restart2Params(monkeypatch):
         lines = f.readlines()
     os.remove("fb.log")
     rmtree("forcefield")
+    cleanOptDir(".")
     os.chdir("..")
     assert len(lines) == 3
 
@@ -977,6 +982,7 @@ def test_restart3Params(monkeypatch):
         lines = f.readlines()
     os.remove("fb.log")
     rmtree("forcefield")
+    cleanOptDir(".")
     os.chdir("..")
     assert len(lines) == 2
 
@@ -1009,5 +1015,6 @@ def test_restart3Params(monkeypatch):
         lines = f.readlines()
     os.remove("fb.log")
     rmtree("forcefield")
+    cleanOptDir(".")
     os.chdir("..")
     assert len(lines) == 1
